requirements.txt
numpy
scipy
torchvision
logging
opencv-python
tqdm
blobfile
pyYaml
pillow